python_api/requirements.txt
fastapi
uvicorn
msgpack
python-multipart
PyGuitarPro